+++ api/src/shared/application/country/country-controller.js
@@ -1,8 +1,8 @@
-import { usecases } from '../domain/usecases/index.js';
-import * as countrySerializer from '../infrastructure/serializers/country-serializer.js';
+import { sharedUsecases } from '../../domain/usecases/index.js';
+import * as countrySerializer from '../../infrastructure/serializers/jsonapi/country-serializer.js';
 
 const findCountries = async function (_request, _h, dependencies = { countrySerializer }) {
-  const countries = await usecases.findCountries();
+  const countries = await sharedUsecases.findCountries();
   return dependencies.countrySerializer.serialize(countries);
 };
