+++ api/src/shared/domain/usecases/index.js
@@ -26,12 +26,14 @@ const dependencies = {
 };
 
 import { deleteUnassociatedBadge } from './delete-unassociated-badge.js';
+import { findCountries } from './find-countries.js';
 import { getAssessment } from './get-assessment.js';
 import { updateAssessmentWithNextChallenge } from './update-assessment-with-next-challenge.js';
 import { updateLastQuestionState } from './update-last-question-state.js';
 
 const usecasesWithoutInjectedDependencies = {
   deleteUnassociatedBadge,
+  findCountries,
   getAssessment,
   updateAssessmentWithNextChallenge,
   updateLastQuestionState,
