+++ api/tests/certification/enrolment/unit/application/country-route_test.js
@@ -1,5 +1,5 @@
-import { countryController } from '../../../../../src/certification/enrolment/application/country-controller.js';
 import * as moduleUnderTest from '../../../../../src/certification/enrolment/application/country-route.js';
+import { countryController } from '../../../../../src/shared/application/country/country-controller.js';
 import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 
 describe('Certification | Enrolment | Unit | Application | country-route', function () {
